Former-commit-id: 6c6ee85e2f71e4f5e75ec5fc20217c0ad37d4a29
Former-commit-id: 07aed6945e82ab7196a85c6bc04f4a40344ed278

+++ registration-processor/stages/packet-receiver-stage/src/main/java/io/mosip/registration/processor/packet/receiver/stage/PacketReceiverStage.java
@@ -146,8 +146,11 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
 			throw new UnexpectedException(e.getMessage());
 		} finally {
-			if (file.exists())
-				deleteFile(file);
+			if(file != null) {
+				if (file.exists()) {
+					deleteFile(file);
+				}
+			}
 		}
 	}
